+++ template/UtilisateurList.php
@@ -3,7 +3,7 @@
 /**
  * @var Gabarit $this
  * @var bool $droitCreation
- * @var bool $droitEdition
+ * @var bool $droitSuppression
  * @var string $descendance
  * @var array $all_role
  * @var string $role_selected
@@ -12,6 +12,7 @@
  * @var array $liste_utilisateur
  * @var int $id_e
  * @var int $offset
+ * @var int $id_u_courant
  */
 
 $exportUserUrl = sprintf(
@@ -73,6 +74,12 @@ $exportUserUrl = sprintf(
    href='<?php hecho($exportUserUrl); ?>'
 ><i class='fas fa-download'></i>&nbsp;Exporter</a>
 
+<?php if ($droitSuppression) : ?>
+    <button type='submit' form='form-suppression-lot' class='btn btn-danger' id='btn-suppression-lot-top' disabled>
+        <i class='fa fa-trash'></i>&nbsp;Supprimer la sélection
+    </button>
+<?php endif; ?>
+
     <?php
     $this->suivantPrecedent(
         $offset,
@@ -81,7 +88,7 @@ $exportUserUrl = sprintf(
         "Entite/utilisateur?id_e=$id_e&page=1&search=$search&descendance=$descendance&role_selected=$role_selected"
     ); ?>
 
-<?php if ($droitEdition) : ?>
+<?php if ($droitSuppression) : ?>
 <form action='Utilisateur/suppression' method='post' id='form-suppression-lot'>
     <?php $this->displayCSRFInput() ?>
     <input type='hidden' name='id_e' value='<?= $id_e ?>'/>
@@ -94,7 +101,7 @@ $exportUserUrl = sprintf(
 <table class='table table-striped'>
 <thead>
 <tr>
-    <?php if ($droitEdition) : ?>
+    <?php if ($droitSuppression) : ?>
         <th><input type='checkbox' id='select-all-users' title='Tout sélectionner'/></th>
     <?php endif; ?>
     <th class='w200'>Prénom Nom</th>
@@ -110,9 +117,11 @@ $exportUserUrl = sprintf(
 
 <?php foreach ($liste_utilisateur as $user) : ?>
     <tr>
-        <?php if ($droitEdition) : ?>
+        <?php if ($droitSuppression) : ?>
             <td>
-                <input type='checkbox' name='id_u_list[]' value='<?= $user['id_u'] ?>' class='user-checkbox'/>
+                <?php if ($user['id_u'] !== $id_u_courant) : ?>
+                    <input type='checkbox' name='id_u_list[]' value='<?= $user['id_u'] ?>' class='user-checkbox'/>
+                <?php endif; ?>
             </td>
         <?php endif; ?>
         <td>
@@ -186,7 +195,7 @@ $exportUserUrl = sprintf(
        href='<?php hecho($exportUserUrl); ?>'
     ><i class='fas fa-download'></i>&nbsp;Exporter</a>
 
-<?php if ($droitEdition) : ?>
+<?php if ($droitSuppression) : ?>
     <button type='submit' form='form-suppression-lot' class='btn btn-danger' id='btn-suppression-lot' disabled>
         <i class='fa fa-trash'></i>&nbsp;Supprimer la sélection
     </button>
@@ -200,6 +209,7 @@ $exportUserUrl = sprintf(
     function updateDeleteButton() {
         const anyChecked = document.querySelectorAll('.user-checkbox:checked').length > 0;
         document.getElementById('btn-suppression-lot').disabled = !anyChecked;
+        document.getElementById('btn-suppression-lot-top').disabled = !anyChecked;
     }
 </script>
 <?php endif; ?>